requirements.txt
matplotlib==3.7.1
pandas==1.5.3
pandasai==0.5.0
python-dotenv==1.0.0
streamlit==1.23.1
